+++ src/main/java/com/limechain/network/protocol/grandpa/messages/commit/Vote.java
@@ -3,8 +3,10 @@ package com.limechain.network.protocol.grandpa.messages.commit;
 import io.emeraldpay.polkaj.types.Hash256;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class Vote {
     private Hash256 blockHash;
-    private long blockNumber;
+    private BigInteger blockNumber;
 }
